Introduce ActiveRecord::Base.find
This commit introduces the find method to our `ActiveRecord::Base`
class. The method expects an open connection to the db and simply relies
on the `SqliteAdapter.execute` method to run the SQL query. The first
result containing the attributes is selected and a new Post object is
returned using the attributes above.