Introduce `ActiveRecord::Base.all`
This commit introduces the `.all` method to our `ActiveRecord::Base`
class. The method follows the footsteps of the `.find` method introduced
earlier. The difference simply lies in the SQL query issued and also
what the returned. We select all posts from the db, and return an array
of `Post` objects.