Implement the Relations class.
* Why was this change necessary?
This commit introduces the ability to query using the `where` keyword.
`where` takes a string which must be a valid sql statement. We are also
able to chain multiple `where` commands.

* How does it address the problem?
A new `Relations` class was introduced to implement this feature. The
new class keeps track of the where commands using an array. The `to_sql`
method builds the full sql statement using all the commands. Finally,
`records` fetches all records from the db by executing the sql command.
The `fist`. `each` and `length` methods were also introduced which
simply delegate to the `records` object.